Skip Steam installation check when -nosteam flag is passed

+++ src/client/main.cpp
@@ -44,12 +44,15 @@ namespace
 
 	bool restart_app_if_necessary_stub()
 	{
-		const std::string steam_path = steam::SteamAPI_GetSteamInstallPath();
-		if (steam_path.empty() || !utils::io::file_exists(steam_path + "/steam.exe"))
+		if (!utils::flags::has_flag("nosteam"))
 		{
-			game::show_error("Steam must be installed for the game to run. Please install Steam!");
-			ShellExecuteA(nullptr, "open", "https://store.steampowered.com/about/", nullptr, nullptr, SW_SHOWNORMAL);
-			TerminateProcess(GetCurrentProcess(), 1);
+			const std::string steam_path = steam::SteamAPI_GetSteamInstallPath();
+			if (steam_path.empty() || !utils::io::file_exists(steam_path + "/steam.exe"))
+			{
+				game::show_error("Steam must be installed for the game to run. Please install Steam!");
+				ShellExecuteA(nullptr, "open", "https://store.steampowered.com/about/", nullptr, nullptr, SW_SHOWNORMAL);
+				TerminateProcess(GetCurrentProcess(), 1);
+			}
 		}
 
 		utils::hook::set(g_original_import.first, g_original_import.second);
